+++ src/test/java/com/vaadin/ui/combobox/test/ComboBoxPageIT.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.combobox;
+package com.vaadin.ui.combobox.test;
 
 import java.util.List;
 import java.util.Map;
