+++ src/test/java/com/vaadin/ui/combobox/test/ComboBoxTest.java
@@ -13,12 +13,14 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.combobox;
+package com.vaadin.ui.combobox.test;
 
 import java.util.Arrays;
 
 import org.junit.Assert;
 import org.junit.Test;
+
+import com.vaadin.ui.combobox.ComboBox;
 
 import elemental.json.JsonArray;
 import elemental.json.JsonObject;
